split cbhdl
